refactor: BaseInput 생성 및 적용
