+++ src/components/pages/CardRegister/CardExpirationDateInput/CardExpirationDateInput.styles.tsx
@@ -1,21 +1,11 @@
-import styled, { DefaultTheme } from 'styled-components';
+import styled from "styled-components";
+import { BaseInput } from "../../../@common/Input/InputStyles.styles";
 
-interface StyledInputProps {
-  inputTheme?: keyof DefaultTheme;
-}
-
-export const Input = styled.input<StyledInputProps>`
+export const Input = styled.input`
+  ${BaseInput}
   width: 30%;
-  height: 100%;
-  font-size: 12px;
   text-align: center;
   padding: 0;
-  border: none;
-  outline: none;
-  background-color: ${(props) =>
-    props.inputTheme
-      ? props.theme[props.inputTheme].backgroundColor
-      : props.theme.pale.backgroundColor};
 `;
 
 export const FieldSet = styled.fieldset`
