+++ src/components/pages/CardRegister/CardNameInput/CardNameInput.styles.tsx
@@ -1,27 +1,13 @@
-import styled, { DefaultTheme } from 'styled-components';
+import styled, { DefaultTheme } from "styled-components";
+import { BaseInput } from "../../../@common/Input/InputStyles.styles";
 
 interface StyledInputProps {
   inputTheme?: keyof DefaultTheme;
 }
 
 export const Input = styled.input<StyledInputProps>`
-  width: 100%;
-  height: 100%;
-  font-size: 12px;
-  padding: 0;
-  border: none;
-  border-radius: 7px;
-  outline: none;
+  ${BaseInput}
   padding: 0 20px;
-  background-color: ${(props) =>
-    props.inputTheme
-      ? props.theme[props.inputTheme].backgroundColor
-      : props.theme.pale.backgroundColor};
-    };
-
-  &:-webkit-autofill {
-    -webkit-box-shadow: 0 0 0 1000px #ecebf1 inset !important;
-  }
 `;
 
 export const Label = styled.label`
